Update documentation and enable debugging in vscode
Updated the documentation in the README to describe how to build and
debug within vscode, and also added tasks.json and launch.json files to
allow debugging within vscode during development.

In addition, a comment was added in the types file to improve
understanding of the type_mgr::bind method.